get section attribute by id

See merge request tike-ohtu/orgrek-ou-service!79

+++ src/main/java/fi/helsinki/ohtu/orgrekouservice/controller/SectionController.java
@@ -2,13 +2,16 @@ package fi.helsinki.ohtu.orgrekouservice.controller;
 
 import fi.helsinki.ohtu.orgrekouservice.domain.EmptyJsonResponse;
 import fi.helsinki.ohtu.orgrekouservice.domain.SectionAttribute;
+import fi.helsinki.ohtu.orgrekouservice.service.HierarchyFilterService;
 import fi.helsinki.ohtu.orgrekouservice.service.SectionAttributeService;
+import fi.helsinki.ohtu.orgrekouservice.service.SectionValidationService;
+import fi.helsinki.ohtu.orgrekouservice.util.Constants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -17,6 +20,12 @@ public class SectionController {
 
     @Autowired
     private SectionAttributeService sectionAttributeService;
+
+    @Autowired
+    private HierarchyFilterService hierarchyFilterService;
+
+    @Autowired
+    private SectionValidationService sectionValidationService;
 
     @GetMapping("/all")
     public ResponseEntity<List<SectionAttribute>> getSectionAttributes() {
@@ -27,8 +36,15 @@ public class SectionController {
     @PostMapping("/insert")
     public ResponseEntity insertSectionAttribute(@RequestBody SectionAttribute sectionAttribute) {
         try {
-            SectionAttribute insertedSectionAttribute = sectionAttributeService.insertSectionAttribute(sectionAttribute);
-            return new ResponseEntity<>(insertedSectionAttribute, HttpStatus.OK);
+            List<String> distinctHierarchyFilterKeys = hierarchyFilterService.getDistinctHierarchyFilterKeys();
+            List<SectionAttribute> sectionAttributeList = sectionAttributeService.getAllSectionAttributes();
+            ResponseEntity response = sectionValidationService.validateSectionAttributes(distinctHierarchyFilterKeys, sectionAttribute, Constants.NEW_SECTION_ATTRIBUTE, sectionAttributeList);
+            if (response.getStatusCode().equals(HttpStatus.OK)) {
+                SectionAttribute insertedSectionAttribute = sectionAttributeService.insertSectionAttribute(sectionAttribute);
+                return new ResponseEntity<>(insertedSectionAttribute, HttpStatus.OK);
+            } else {
+                return new ResponseEntity(response.getBody(), response.getStatusCode());
+            }
         } catch (Exception e) {
             return new ResponseEntity<>(new EmptyJsonResponse(),HttpStatus.BAD_REQUEST);
         }
@@ -37,8 +53,17 @@ public class SectionController {
     @PutMapping("/update")
     public ResponseEntity updateSectionAttribute(@RequestBody SectionAttribute sectionAttribute) {
         try {
-            sectionAttributeService.updateSectionAttribute(sectionAttribute);
-            return new ResponseEntity<>(new EmptyJsonResponse(), HttpStatus.OK);
+            List<String> distinctHierarchyFilterKeys = hierarchyFilterService.getDistinctHierarchyFilterKeys();
+            SectionAttribute foundSectionAttribute = sectionAttributeService.getSectionAttributeById(sectionAttribute.getId());
+            SectionAttribute updatedSectionAttribute = sectionAttributeService.updateSectionAttributeInfo(foundSectionAttribute, sectionAttribute);
+            List<SectionAttribute> sectionAttributeList = sectionAttributeService.getAllSectionAttributes();
+            ResponseEntity response = sectionValidationService.validateSectionAttributes(distinctHierarchyFilterKeys, updatedSectionAttribute, Constants.UPDATE_SECTION_ATTRIBUTE, sectionAttributeList);
+            if (response.getStatusCode().equals(HttpStatus.OK)) {
+                sectionAttributeService.updateSectionAttribute(sectionAttribute);
+                return new ResponseEntity<>(new EmptyJsonResponse(), HttpStatus.OK);
+            } else {
+                return new ResponseEntity(response.getBody(), response.getStatusCode());
+            }
         } catch (Exception e) {
             return new ResponseEntity<>(new EmptyJsonResponse(),HttpStatus.BAD_REQUEST);
         }
@@ -47,8 +72,16 @@ public class SectionController {
     @DeleteMapping("/{id}/delete")
     public ResponseEntity deleteSectionAttribute(@PathVariable("id") int sectionId) {
         try {
-            HttpStatus httpStatus = sectionAttributeService.deleteSectionAttribute(sectionId);
-            return new ResponseEntity<>(new EmptyJsonResponse(), httpStatus);
+            List<String> distinctHierarchyFilterKeys = hierarchyFilterService.getDistinctHierarchyFilterKeys();
+            List<SectionAttribute> sectionAttributeList = sectionAttributeService.getAllSectionAttributes();
+            SectionAttribute foundSectionAttribute = sectionAttributeService.getSectionAttributeById(sectionId);
+            ResponseEntity response = sectionValidationService.validateSectionAttributes(distinctHierarchyFilterKeys, foundSectionAttribute, Constants.DELETE_SECTION_ATTRIBUTE, sectionAttributeList);
+            if (response.getStatusCode().equals(HttpStatus.OK)) {
+                HttpStatus httpStatus = sectionAttributeService.deleteSectionAttribute(sectionId);
+                return new ResponseEntity<>(new EmptyJsonResponse(), httpStatus);
+            }  else {
+                return new ResponseEntity(response.getBody(), response.getStatusCode());
+            }
         } catch (Exception e) {
             return new ResponseEntity<>(new EmptyJsonResponse(),HttpStatus.BAD_REQUEST);
         }
